Corrección de avisso
Se pusieron los avisos y modales de los días festivos

+++ src/app/modules/config/festive-days/festive-days.component.ts
@@ -1,5 +1,7 @@
 import { Component, OnInit } from '@angular/core';
+import { MatSnackBar, MatSnackBarVerticalPosition } from '@angular/material/snack-bar';
 import { ApiService } from 'src/app/shared/services/api.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-festive-days',
@@ -9,7 +11,8 @@ import { ApiService } from 'src/app/shared/services/api.service';
 export class FestiveDaysComponent implements OnInit {
 
   dataSource = [{id:1, fecha:20, nombre:'navida', estatus:1}]
-
+  verticalPosition: MatSnackBarVerticalPosition = 'top';
+  config_snack = { duration: 3000,verticalPosition: this.verticalPosition}
   day:Date
   name:String
 
@@ -20,7 +23,7 @@ export class FestiveDaysComponent implements OnInit {
     dia: 'Dia'
   };
 
-  constructor( public apiService:ApiService, ) { }
+  constructor( public apiService:ApiService, public snackBar: MatSnackBar ) { }
 
   ngOnInit(): void {
     this.getDays()
@@ -38,6 +41,10 @@ export class FestiveDaysComponent implements OnInit {
         console.log(res)
         //res = JSON.parse(this.apiService.decrypt(res.message,this.apiService.getPrivateKey()))
         this.getDays()
+        this.snackBar.open('Día festivo agregado exitosamente', '', this.config_snack);
+      },
+      error: err => {
+        this.snackBar.open('Ocurrió un error al agregar el día festivo', '', this.config_snack);
       }
     })
   }
@@ -48,6 +55,9 @@ export class FestiveDaysComponent implements OnInit {
         console.log(res)
         //res = JSON.parse(this.apiService.decrypt(res.message,this.apiService.getPrivateKey()))
         this.dataSource = res.result
+      },
+      error: err => {
+        this.snackBar.open('Ocurrió un error al obtener los días festivos', '', this.config_snack);
       }
     })
   }
@@ -56,12 +66,27 @@ export class FestiveDaysComponent implements OnInit {
     const body = {
       id: element.id
     }
-    this.apiService.deleteDiaFestivo(body).subscribe({
-      next:res => {
-        console.log(res)
-        //res = JSON.parse(this.apiService.decrypt(res.message,this.apiService.getPrivateKey()))
-        this.getDays()
-      }
-    })
+
+    Swal.fire({
+      title: "¿Seguro que desea borrar el día festivo?",
+      showDenyButton: true,
+      showCancelButton: true,
+      showConfirmButton:false,
+      denyButtonText: `Borrar`,
+      cancelButtonText:'Cancelar'
+    }).then((result) => {
+      /* Read more about isConfirmed, isDenied below */
+      if (result.isDenied) {
+        this.apiService.deleteDiaFestivo(body).subscribe({
+          next:res => {
+            console.log(res)
+            //res = JSON.parse(this.apiService.decrypt(res.message,this.apiService.getPrivateKey()))
+            this.getDays()
+            this.snackBar.open('Día festivo borrado exitosamente', '', this.config_snack);
+          }
+        })
+      } 
+    });
+    
   }
 }
